Module egui_dock::dock_state

The main structure of the library.

Re-exports

pub use surface::Surface;

pub use surface_index::SurfaceIndex;

pub use window_state::WindowState;

Modules

surface

Represents an area in which a dock tree is rendered.

surface_index

Wrapper around indices to the collection of surfaces inside a DockState.

translations

Specifies text displayed in different elements of the DockArea.

tree

Binary tree representing the relationships between Nodes.

window_state

Window states which tells floating tabs how to be displayed inside their window,

Structs

DockState

The heart of egui_dock.