egui_dock::dock_state::tree::node

Struct LeafNode

pub struct LeafNode<Tab> {
    pub rect: Rect,
    pub viewport: Rect,
    pub tabs: Vec<Tab>,
    pub active: TabIndex,
    pub scroll: f32,
    pub collapsed: bool,
}

The inner data of a Node::Leaf, which contains tabs and can be collapsed.

Fields
rect: Rect

The full rectangle - tab bar plus tab body.

viewport: Rect

The tab body rectangle.

tabs: Vec<Tab>

All the tabs in this node.

active: TabIndex

The opened tab.

scroll: f32

Scroll amount of the tab bar.

collapsed: bool

Whether the leaf is collapsed.

Implementations

impl<Tab> LeafNode<Tab>

pub const fn new(tabs: Vec<Tab>) -> Self

Create New LeafNode with specified tabs, all other internal values will be filled by โnothingโ defaults.

pub fn set_active_tab(&mut self, active_tab: impl Into<TabIndex>)

Set the active tab of this LeafNode

If active_tab is out of bounds, it will be ignored and the active tab will not be changed.

pub fn set_rect(&mut self, new_rect: Rect)

Set the area this LeafNode Occupies on screen.

pub fn len(&self) -> usize

Get the length of tab list in this LeafNode.

pub fn is_empty(&self) -> bool

Returns true when the LeafNode contains no tabs.

pub fn rect(&self) -> Rect

Get a [Rect] representing the area this LeafNode occupies on screen.

pub fn tabs(&self) -> &[Tab]

Get immutable access to the Tabs of this LeafNode

pub fn tabs_mut(&mut self) -> &mut [Tab]

Get mutable access to the Tabs of this LeafNode

pub fn append_tab(&mut self, tab: Tab)

Append a Tab to the end of this LeafNodes tab list.

This will also focus the added tab.

pub fn insert_tab(&mut self, tab_index: impl Into<TabIndex>, tab: Tab)

Insert a Tab to this LeafNodes tab list at the specified TabIndex.

This will also focus the added tab.

Panics

if tab_index exceeds the leafโs tab list length.

pub fn remove_tab(&mut self, tab_index: impl Into<TabIndex>) -> Option<Tab>

Remove a Tab to this LeafNodes tab list at the specified TabIndex.

This will also focus the added tab.โ

Panics

if tab_index is out of bounds for the tab list

pub fn retain_tabs<F>(&mut self, predicate: F)
where F: FnMut(&mut Tab) -> bool,

Removes all tabs for which predicate returns false.

pub fn active_focused(&mut self) -> Option<(Rect, &mut Tab)>

Return the area and tab which is currently representing this LeafNode

This may return None if the leaf contains 0 tabs.