Module egui_dock::style

Look and feel.

Structs

ButtonsStyle

Specifies the look and feel of buttons.

LeafHighlighting

Highlighting on the currently hovered leaf.

OverlayFeel

Specifies the feel of the tab drop overlay, i.e anything non visual about the overlay.

OverlayStyle

Specifies the look and feel of the tab drop overlay.

SeparatorStyle

Specifies the look and feel of node separators.

Style

Lets you change how tabs and the DockArea should look and feel. Style is divided into several, more specialized structs that handle individual elements of the UI.

TabBarStyle

Specifies the look and feel of tab bars.

TabBodyStyle

Specifies the look and feel of the tab body.

TabInteractionStyle

Specifies the look and feel of individual tabs while they are being interacted with.

TabStyle

Specifies the look and feel of an individual tab.

Enums

OverlayType

Specifies the type of overlay used.

TabAddAlign

Left or right alignment for tab add button.