Struct egui_dock::style::TabBarStyle

pub struct TabBarStyle {
    pub bg_fill: Color32,
    pub height: f32,
    pub show_scroll_bar_on_overflow: bool,
    pub rounding: Rounding,
    pub hline_color: Color32,
    pub fill_tab_bar: bool,
}

Specifies the look and feel of tab bars.

Fields
bg_fill: Color32

Background color of tab bar. By Default it’s Color32::WHITE.

height: f32

Height of the tab bar. By Default it’s 24.0.

show_scroll_bar_on_overflow: bool

Show a scroll bar when tab bar overflows. By Default it’s true.

rounding: Rounding

Tab rounding. By Default it’s Rounding::default.

hline_color: Color32

Color of th line separating the tab name area from the tab content area. By Default it’s Color32::BLACK.

fill_tab_bar: bool

Whether tab titles expand to fill the width of their tab bars.

Implementations

impl TabBarStyle

pub fn from_egui(style: &Style) -> Self

Derives relevant fields from egui::Style and sets the remaining fields to their default values.

Fields overwritten by egui::Style are:

TabBarStyle::bg_fill

TabBarStyle::rounding

TabBarStyle::hline_color