Module egui_dock::dock_state::tree

Binary tree representing the relationships between Nodes.

Implementation details

The binary tree is stored in a Vec indexed by NodeIndex. The root is always at index 0. For a given node n:

left child of n will be at index n * 2 + 1.

right child of n will be at index n * 2 + 2.

Re-exports

pub use node::Node;

pub use node_index::NodeIndex;

pub use tab_index::TabIndex;

pub use tab_iter::TabIter;

Modules

node

Represents an abstract node of a Tree.

node_index

Wrapper around indices to the collection of nodes inside a Tree.

tab_index

Identifies a tab within a Node.

tab_iter

Iterates over all tabs in a Tree.

Structs

Tree

Binary tree representing the relationships between Nodes.

Enums

Split

Direction in which a new node is created relatively to the parent node at which the split occurs.

TabDestination

The destination for a tab which is being moved.

TabInsert

Specify how a tab should be added to a Node.