egui_dock::widgets::tab_viewer

Trait TabViewer

pub trait TabViewer {
    type Tab;

    // Required methods
    fn title(&mut self, tab: &mut Self::Tab) -> WidgetText;
    fn ui(&mut self, ui: &mut Ui, tab: &mut Self::Tab);

    // Provided methods
    fn context_menu(
        &mut self,
        _ui: &mut Ui,
        _tab: &mut Self::Tab,
        _surface: SurfaceIndex,
        _node: NodeIndex,
    ) { ... }
    fn id(&mut self, tab: &mut Self::Tab) -> Id { ... }
    fn on_tab_button(&mut self, _tab: &mut Self::Tab, _response: &Response) { ... }
    fn on_close(&mut self, _tab: &mut Self::Tab) -> OnCloseResponse { ... }
    fn is_closeable(&self, _tab: &Self::Tab) -> bool { ... }
    fn closeable(&mut self, _tab: &mut Self::Tab) -> bool { ... }
    fn force_close(&mut self, _tab: &mut Self::Tab) -> bool { ... }
    fn on_add(&mut self, _surface: SurfaceIndex, _node: NodeIndex) { ... }
    fn on_rect_changed(&mut self, _tab: &mut Self::Tab) { ... }
    fn add_popup(
        &mut self,
        _ui: &mut Ui,
        _surface: SurfaceIndex,
        _node: NodeIndex,
    ) { ... }
    fn tab_style_override(
        &self,
        _tab: &Self::Tab,
        _global_style: &TabStyle,
    ) -> Option<TabStyle> { ... }
    fn allowed_in_windows(&self, _tab: &mut Self::Tab) -> bool { ... }
    fn clear_background(&self, _tab: &Self::Tab) -> bool { ... }
    fn scroll_bars(&self, _tab: &Self::Tab) -> [bool; 2] { ... }

}

Defines how a tab should behave and be rendered inside a Tree.

Required Associated Types

type Tab

The type of tab in which you can store state to be drawn in your tabs.

Required Methods

fn title(&mut self, tab: &mut Self::Tab) -> WidgetText

The title to be displayed in the tab bar.

fn ui(&mut self, ui: &mut Ui, tab: &mut Self::Tab)

Actual tab content.

Provided Methods

fn context_menu( &mut self, _ui: &mut Ui, _tab: &mut Self::Tab, _surface: SurfaceIndex, _node: NodeIndex, )

Content inside the context menu shown when the tab is right-clicked.

_surface and _node specify which Surface and Node that this particular context menu belongs to.

fn id(&mut self, tab: &mut Self::Tab) -> Id

Unique ID for this tab.

If not implemented, uses tab title text as an ID source.

fn on_tab_button(&mut self, _tab: &mut Self::Tab, _response: &Response)

Called after each tab button is shown, so you can add a tooltip, check for clicks, etc.

fn on_close(&mut self, _tab: &mut Self::Tab) -> OnCloseResponse

This is called when the _tab gets closed by the user.

Returns an OnCloseResponse which determines what happens to the tab after this function gets called.

fn is_closeable(&self, _tab: &Self::Tab) -> bool

Returns true if the user of your app should be able to close a given _tab.

By default, true is always returned.

fn closeable(&mut self, _tab: &mut Self::Tab) -> bool

👎Deprecated: Use the TabViewer::is_closeable function instead.

Returns true if the user of your app should be able to close a given _tab.

By default, true is always returned.

fn force_close(&mut self, _tab: &mut Self::Tab) -> bool

This is called every frame after ui is called, if the _tab is active.

Returns true if the tab should be forced to close, false otherwise.

In the event this function returns true the tab will be removed without calling on_close.

fn on_add(&mut self, _surface: SurfaceIndex, _node: NodeIndex)

This is called when the add button is pressed.

_surface and _node specify which Surface and on which Node this particular add button was pressed.

fn on_rect_changed(&mut self, _tab: &mut Self::Tab)

Called when the rectangle of the tab content changes.

This can happen when the window is resized, panels are docked or undocked, or when the layout of the dock area is changed in any way that affects the available space for the tab content.

This is useful for tabs that need to adjust their content based on the available space.

fn add_popup(&mut self, _ui: &mut Ui, _surface: SurfaceIndex, _node: NodeIndex)

Content of the popup under the add button. Useful for selecting what type of tab to add.

This requires that DockArea::show_add_buttons and DockArea::show_add_popup are set to true.

fn tab_style_override( &self, _tab: &Self::Tab, _global_style: &TabStyle, ) -> Option<TabStyle>

Sets custom style for given tab.

fn allowed_in_windows(&self, _tab: &mut Self::Tab) -> bool

Specifies a tab’s ability to be shown in a window.

Returns false if this tab should never be turned into a window.

fn clear_background(&self, _tab: &Self::Tab) -> bool

Whether the tab body will be cleared with the color specified in TabBarStyle::bg_fill.

fn scroll_bars(&self, _tab: &Self::Tab) -> [bool; 2]

Returns true if the horizontal and vertical scroll bars will be shown for tab.

By default, both scroll bars are shown.